requirements.txt
adafruit-circuitpython-ssd1306==2.12.10
pillow==9.5.0 
numpy==1.24.0
adafruit-blinka==8.4.2
psutil==5.9.2
python-dateutil==2.8.2
qrcode[pil]==7.4.2